Move all my decorator design code into another file
